Massive reorganisation of project structure to bring scene tscn files next to their Kt scripts. Broke the project for a while but seems to be behaving now.
