+++ gdj/state/GameState.gdj
@@ -2,8 +2,8 @@
 // Note: You can however freely move this file inside your godot project if you want. Keep in mind however, that if you rename the originating source code file, this file will be deleted and regenerated as a new file instead of being updated! Other modifications to the source file however, will result in this file being updated.
 
 registeredName = GameState
-fqName = org.liamjd.kuiper.state.GameState
-relativeSourcePath = src/main/kotlin/org/liamjd/kuiper/state/GameState.kt
+fqName = state.GameState
+relativeSourcePath = src/main/kuiper/state/GameState.kt
 baseType = Node
 supertypes = [
     godot.Node,
