+++ src/main/kuiper/screens/kuiper/escMenu/EscMenu.kt
@@ -1,0 +1,25 @@
+package screens.Kuiper.EscMenu
+
+import godot.Control
+import godot.annotation.RegisterClass
+import godot.annotation.RegisterFunction
+
+@RegisterClass
+class EscMenu: Control() {
+
+	// Declare member variables here. Examples:
+	// val a = 2
+	// val b = "text"
+
+	// Called when the node enters the scene tree for the first time.
+	@RegisterFunction
+	override fun _ready() {
+		
+	}
+
+	// Called every frame. 'delta' is the elapsed time since the previous frame.
+	@RegisterFunction
+	override fun _process(delta: Double) {
+		
+	}
+}
