+++ src/main/kuiper/screens/kuiper/game.tscn
@@ -1,0 +1,114 @@
+[gd_scene load_steps=2 format=3 uid="uid://dhc5at6r5jl4n"]
+
+[ext_resource type="Script" path="res://src/main/kuiper/screens/Kuiper/KuiperGame.kt" id="1_nlsq8"]
+
+[node name="CanvasLayer" type="CanvasLayer"]
+script = ExtResource("1_nlsq8")
+
+[node name="Background" type="ColorRect" parent="."]
+anchors_preset = 15
+anchor_right = 1.0
+anchor_bottom = 1.0
+grow_horizontal = 2
+grow_vertical = 2
+color = Color(0.125911, 0.125911, 0.125911, 1)
+
+[node name="AspectRatioContainer" type="MarginContainer" parent="Background"]
+layout_mode = 1
+anchors_preset = 15
+anchor_right = 1.0
+anchor_bottom = 1.0
+grow_horizontal = 2
+grow_vertical = 2
+theme_override_constants/margin_left = 10
+theme_override_constants/margin_top = 10
+theme_override_constants/margin_right = 10
+theme_override_constants/margin_bottom = 10
+
+[node name="VBoxContainer" type="VBoxContainer" parent="Background/AspectRatioContainer"]
+layout_mode = 2
+
+[node name="TopRow_hbox" type="HBoxContainer" parent="Background/AspectRatioContainer/VBoxContainer"]
+layout_mode = 2
+alignment = 1
+
+[node name="PanelContainer" type="PanelContainer" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox"]
+layout_mode = 2
+
+[node name="MenuButton" type="Button" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox/PanelContainer"]
+layout_mode = 2
+
+[node name="MarginContainer" type="MarginContainer" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox"]
+layout_mode = 2
+theme_override_constants/margin_left = 120
+
+[node name="ResourcesPanel" type="PanelContainer" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox"]
+layout_mode = 2
+
+[node name="HBoxContainer" type="HBoxContainer" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox/ResourcesPanel"]
+layout_mode = 2
+
+[node name="Label" type="Label" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox/ResourcesPanel/HBoxContainer"]
+layout_mode = 2
+text = "Influence"
+
+[node name="Label2" type="Label" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox/ResourcesPanel/HBoxContainer"]
+layout_mode = 2
+text = "Money"
+
+[node name="Label" type="Label" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox"]
+layout_mode = 2
+size_flags_horizontal = 3
+theme_override_colors/font_color = Color(0.952941, 0.509804, 0.0705882, 1)
+theme_override_font_sizes/font_size = 20
+text = "Project Kuiper"
+horizontal_alignment = 1
+
+[node name="ResearchPanel" type="PanelContainer" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox"]
+layout_mode = 2
+
+[node name="HBoxContainer" type="HBoxContainer" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox/ResearchPanel"]
+layout_mode = 2
+
+[node name="Label" type="Label" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox/ResearchPanel/HBoxContainer"]
+layout_mode = 2
+text = "Physics"
+
+[node name="Label2" type="Label" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox/ResearchPanel/HBoxContainer"]
+layout_mode = 2
+text = "Biology
+"
+
+[node name="MarginContainer2" type="MarginContainer" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox"]
+layout_mode = 2
+theme_override_constants/margin_left = 120
+
+[node name="PanelContainer2" type="PanelContainer" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox"]
+layout_mode = 2
+
+[node name="Year_lbl" type="Label" parent="Background/AspectRatioContainer/VBoxContainer/TopRow_hbox/PanelContainer2"]
+layout_mode = 2
+size_flags_horizontal = 8
+text = "Year"
+
+[node name="TabBar" type="TabBar" parent="Background/AspectRatioContainer/VBoxContainer"]
+layout_mode = 2
+current_tab = 0
+tab_alignment = 1
+tab_count = 3
+tab_0/title = "Earth"
+tab_1/title = "Near Earth Orbit"
+tab_2/title = "Mars"
+
+[node name="MarginContainer" type="MarginContainer" parent="Background/AspectRatioContainer/VBoxContainer"]
+layout_mode = 2
+size_flags_vertical = 3
+
+[node name="BottomRow_hbox" type="HBoxContainer" parent="Background/AspectRatioContainer/VBoxContainer"]
+layout_mode = 2
+size_flags_horizontal = 8
+
+[node name="EndTurn_btn" type="Button" parent="Background/AspectRatioContainer/VBoxContainer/BottomRow_hbox"]
+layout_mode = 2
+size_flags_horizontal = 8
+text = "End Turn"
