+++ src/main/kuiper/screens/main.tscn
@@ -1,0 +1,3 @@
+[gd_scene format=3 uid="uid://e5gl3nsv0vqc"]
+
+[node name="Main" type="Node"]
