+++ src/main/kuiper/screens/mainMenu/main_menu.tscn
@@ -1,0 +1,73 @@
+[gd_scene load_steps=2 format=3 uid="uid://dwd5bpnoghggx"]
+
+[ext_resource type="Script" path="res://src/main/kuiper/screens/MainMenu/MainMenu.kt" id="1_jfeih"]
+
+[node name="MainMenu" type="CanvasLayer"]
+script = ExtResource("1_jfeih")
+
+[node name="ColorRectBackground" type="ColorRect" parent="."]
+anchors_preset = 15
+anchor_right = 1.0
+anchor_bottom = 1.0
+grow_horizontal = 2
+grow_vertical = 2
+color = Color(0.125911, 0.125911, 0.125911, 1)
+
+[node name="VBoxContainer" type="VBoxContainer" parent="."]
+anchors_preset = 8
+anchor_left = 0.5
+anchor_top = 0.5
+anchor_right = 0.5
+anchor_bottom = 0.5
+offset_left = -276.5
+offset_top = -127.0
+offset_right = 276.5
+offset_bottom = 127.0
+grow_horizontal = 2
+grow_vertical = 2
+size_flags_horizontal = 0
+
+[node name="GameTitle" type="Label" parent="VBoxContainer"]
+layout_mode = 2
+theme_override_font_sizes/font_size = 40
+text = "Project Herschel: Heliopause"
+horizontal_alignment = 1
+
+[node name="HSeparator" type="HSeparator" parent="VBoxContainer"]
+custom_minimum_size = Vector2(0, 20)
+layout_mode = 2
+
+[node name="VBoxContainer" type="VBoxContainer" parent="VBoxContainer"]
+custom_minimum_size = Vector2(200, 0)
+layout_mode = 2
+size_flags_horizontal = 4
+size_flags_stretch_ratio = 0.0
+
+[node name="Continue" type="Button" parent="VBoxContainer/VBoxContainer"]
+layout_mode = 2
+size_flags_horizontal = 4
+disabled = true
+text = "Continue"
+
+[node name="NewGame" type="Button" parent="VBoxContainer/VBoxContainer"]
+layout_mode = 2
+text = "New Game"
+
+[node name="Load" type="Button" parent="VBoxContainer/VBoxContainer"]
+layout_mode = 2
+disabled = true
+text = "Load"
+
+[node name="Save" type="Button" parent="VBoxContainer/VBoxContainer"]
+layout_mode = 2
+disabled = true
+text = "Save"
+
+[node name="Quit" type="Button" parent="VBoxContainer/VBoxContainer"]
+layout_mode = 2
+text = "Quit"
+
+[connection signal="new_game_signal" from="." to="." method="_on_new_game"]
+[connection signal="quit_game_signal" from="." to="." method="_on_quit_game"]
+[connection signal="pressed" from="VBoxContainer/VBoxContainer/NewGame" to="." method="_on_new_game"]
+[connection signal="pressed" from="VBoxContainer/VBoxContainer/Quit" to="." method="_on_quit_game"]
